requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.7
black==20.8b1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-graphql-jwt==0.3.1
gevent==20.6.0
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
promise==2.3
psycopg2==2.8.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
supervisor==4.2.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.9
whitenoise==5.1.0
zope.event==4.4
zope.interface==5.1.0
